--- pages/testpage.cpp
@@ -1,5 +1,0 @@
-#include "testpage.h"
-
-TestPage::TestPage(QWidget *parent) {
-
-}
